Use Circle CI machine executor to support Coveralls reporting
Following up from https://github.com/artsy/exchange/pull/270, this
commit updates the Circle CI configuration to use the machine executor
for the test step. This allows us to successfully mount the git
directory to the docker container providing Coveralls with the git
information it needs to upload the test coverage report.